Changes College Data to be loaded in the background so that it desn't take as long to load quicksearch page.
